Load input data from X refactor
