+++ colmap.cpp
@@ -1,0 +1,2 @@
+#include "colmap.hpp"
+
