+++ nerfstudio.hpp
@@ -4,9 +4,8 @@
 #include <iostream>
 #include <string>
 #include <fstream>
-#include <unordered_map>
 #include <torch/torch.h>
-#include <opencv2/calib3d.hpp>
+#include "input_data.hpp"
 #include "vendor/json/json_fwd.hpp"
 
 using json = nlohmann::json;
@@ -40,43 +39,6 @@ namespace ns{
     void to_json(json &j, const Transforms &t);
     void from_json(const json& j, Transforms &t);
 
-    enum CameraType { Perspective };
-    struct Camera{
-        int width = 0;
-        int height = 0;
-        float fx = 0;
-        float fy = 0;
-        float cx = 0;
-        float cy = 0;
-        float k1 = 0;
-        float k2 = 0;
-        float k3 = 0;
-        float p1 = 0;
-        float p2 = 0;
-        torch::Tensor camToWorld;
-        std::string filePath = "";
-        CameraType cameraType = CameraType::Perspective;
-
-        Camera(int width, int height, float fx, float fy, float cx, float cy, 
-            float k1, float k2, float k3, float p1, float p2,
-            const torch::Tensor &camToWorld, const std::string &filePath) : 
-            width(width), height(height), fx(fx), fy(fy), cx(cx), cy(cy), 
-            k1(k1), k2(k2), k3(k3), p1(p1), p2(p2),
-            camToWorld(camToWorld), filePath(filePath) {}
-        
-        torch::Tensor getIntrinsicsMatrix();
-        bool hasDistortionParameters();
-        std::vector<float> undistortionParameters();
-        void scaleOutputResolution(float scaleFactor);
-        torch::Tensor getImage(int downscaleFactor);
-
-        void loadImage(float downscaleFactor);
-        torch::Tensor K;
-        torch::Tensor image;
-
-        std::unordered_map<int, torch::Tensor> imagePyramids;
-    };
-
     Transforms readTransforms(const std::string &filename);
 
     torch::Tensor posesFromTransforms(const Transforms &t);
@@ -84,18 +46,6 @@ namespace ns{
 
     torch::Tensor rotationMatrix(const torch::Tensor &a, const torch::Tensor &b);
 
-    struct Points{
-        torch::Tensor xyz;
-        torch::Tensor rgb;
-    };
-    struct InputData{
-        std::vector<Camera> cameras;
-        float scaleFactor;
-        torch::Tensor transformMatrix;
-        Points points;
-
-        std::tuple<std::vector<Camera>, Camera *> getCameras(bool validate, const std::string &valImage = "random");
-    };
     InputData inputDataFromNerfStudio(const std::string &projectRoot);
 }
